Make grouped arguments first class.
Grouped arguments are onbtained by grouping together all consecutive labeled arguments.
So e.g. (~x,~y, a, ~z) has two groups: (~x,~y) and (~z), and a normal argument a.

Make grouped arguments first class in the codebase.
So they are the lhs of function types, as well as corresponding lhs of function converters.

This enables support of labeled arguments beyond the `make` function in components.