Add @genType.as to specify the name of the field seen from JS.
The @genType.as/@genFlow.as annotation is analogous to @bs.as in bucklescript (https://bucklescript.github.io/docs/en/object).
When applied to a record field, specifies tha name of the field when seen from the JS side.
Typical use is `type_` seen as `type`.
E.g.:

```reason
  [@genType.as "type"]
  type_: string,

```

Since the meaning is clear from the context when applied to a record field, it’s also possibel to omit `.as` and just use `@genType`.

See https://github.com/cristianoc/genFlow/issues/37.