README: new presentation of genType
genType performs a type-directed transformation.
Whether the generated JS code is typed or not, genType changes the runtime representation of the data from Reason types to an idiomatic JS representation.
So the *dynamic* type is affeced. In addition, static type annotations coud *also* be added to the generated code, or not in the case of the untyped back-end.
In any case, the transformation can be described as a type conversion transformation. So I’m thinking the genType name could actually be legit.