Import type definitions from other files.
When a type import is emitted, handle the case where the type is defined in another file.
The type definition must be obtained by the .cmt(i) of that other file.
A cache of type definitions is kept for .cmt(i) files from which at least one file is imported.

Fixes https://github.com/cristianoc/genFlow/issues/43.