Change the grouping of translations for type declarations, and type maps.
Now type declarations compile to declaration, each of which consists of one code item and a list of import types.
So each export statement carries its own list of imports.

Also, the type map carries, for each type in the map, a list of those import types.

In this way it becomes possible, should one decide to inlide type declarations, do also add their required type imports on-demand.

For now, this is a pure refactoring with no behavioural change.