Add support for inlining annotations. Currently off by default.
This means that if a type is annotated wiht @genType, and used another type in its definition, the other type will also be considered as annotated.

This is currently supported within one file, and not across files.

In config. Off by default at the moment.

See “Type Expansion” in https://github.com/cristianoc/genType/issues/70