Support renaming named arguments in function types.
Support renaming of the form `[@genType.as “xRenamed”] (~x) => …`.

There’s a bug in ocaml 4.03.2 where the first annotation is lost.
So it’s not possible to rename the first named argument.
This limitation will go away once the compiler version is upgraded.

See https://github.com/cristianoc/genType/issues/75.