Change the extension of generated files to .regen.js and .regen.tsx.
Test a breaking change to the extension of generated files.

Change the extension of generated fies to make it uniform across TypeScript and Flow, and independent from webpack plugins.

No need for a special webpack plugin to read `.re.js` files instead of `.re` files.

No difference in instructions to specify the current file in `bs.value` with TypeScript of Flow/Untyped.