Use import instead of require in ES modules, excep when values are imported.
Imported values lead to cyclic imports between the generated file and the `.bs.js` file. So in case of imported values, requires are emitted. In other cases, for ES6, imports are equipped.

TypeScript still needs lookint into. Two outstanding issues is e.g. ReasonReact.js is not found. Another issue is that .bs.js files are do not default to “any” anymore, and type errors are given.