Use import instead of require in ES modules, excep when values are im… (#92)
* Use import instead of require in ES modules, excep when values are imported.

Imported values lead to cyclic imports between the generated file and the `.bs.js` file. So in case of imported values, requires are emitted. In other cases, for ES6, imports are equipped.

TypeScript still needs lookint into. Two outstanding issues is e.g. ReasonReact.js is not found. Another issue is that .bs.js files are do not default to “any” anymore, and type errors are given.

* Keep on using require’s for TypeScript

Using import’s for TypeScript with ES6 for untyped modules complicates things. Skipping for now.

* Check that everything works when buckesript emits es6 modules.

* use “es6” instead of “es6-global” for bucklescript compilation