When a renamed type is expoted, export both the original and renamed version.
The renamed version is needed for JS.
The original version is needed when used from other types in the same or other files. While it would be possible to make other references in the same file aware of the renaming, it’s not possible for other files, because of local compilation.

Fixes https://github.com/cristianoc/genType/issues/98.