Work around issue with Flow and function components.
If a function component is type annotated, then it is not checked by Flow.
Work around this issue by defining an auxiliary wrapper function just to call typeof() on it. The wrapper function is not type annotated directly, but its parameters and return are.
This makes Flow type check function components.