Firsrt part: rough logging of output.
So far, log some output on standard output at the end.
Also, there's a global structure in ResolvedName.re, which should really be created from the outside and not be global.
Also, some code should probably be moved out of ResolvedName.re.
Also, it's not clear whether the export should be type annotated, or whether the inference is OK.