Fix issue where `@genType.import` was not supported in interface files.
It looks like `.cmti` are missing in the AST the information necessary to recognize externals: `val_prim` in `value_description`.
In the absence of that information, now we rely on the presence of the `@genType.import` annotation.
The danger is for a non-external to be annotated `@genType.import` by mistake.